chore(admin): convert transfer tokens to TS
