--- packages/core/admin/admin/src/pages/Settings/pages/TransferTokens/EditView/utils/index.js
@@ -1,4 +1,0 @@
-import getDateOfExpiration from './getDateOfExpiration';
-import schema from './schema';
-
-export { getDateOfExpiration, schema };
